# Performance test infrastructure and sample decoding test