--- a/@here/harp-test-utils/lib/ProfileHelper.ts
+++ b/@here/harp-test-utils/lib/ProfileHelper.ts
@@ -631,7 +631,7 @@ let measurePerformanceSamples: PerformanceTestSample[] = [];
 // in Mocha environment log all profile results after each test
 //
 if (typeof after === "function") {
-    afterEach(() => {
+    after(() => {
         reportPerformanceAndReset();
         reportCallCountsAndReset();
     });