--- a/test/performance/package.json
+++ b/test/performance/package.json
@@ -1,0 +1,20 @@
+{
+    "name": "@here/harp-integration-tests",
+    "version": "0.1.0",
+    "private": true,
+    "author": {
+        "name": "HERE Europe B.V.",
+        "url": "https://here.com"
+    },
+    "devDependencies": {
+        "@here/harp-datasource-protocol": "^0.10.0",
+        "@here/harp-geoutils": "^0.10.0",
+        "@here/harp-omv-datasource": "^0.10.0",
+        "@here/harp-mapview": "^0.10.0",
+        "@here/harp-test-utils": "^0.10.0",
+        "@here/harp-examples": "^0.10.0",
+        "@types/chai": "^4.2.2",
+        "chai": "^4.2.0"
+    },
+    "license": "Apache-2.0"
+}