Infrastructure and example test for measuring performance in repeatable way (part of #853 and #852)

* OMVDecoder performance test added. 
* ProfileHelper added

Yes, this is _only_ for nodejs, but CPU-bound performance critical code will behave in similar way as in real browser and if we want to test this in real browsers too, this may be good first step ...

New workflow (also in `README`):

Example usage:
1) `git checkout master`
2)  `PROFILEHELPER_COMMAND=baseline yarn performance-test-node` - create baseline of measurements for your particular platform

3) checkout your branch
4) ~break~ ... improve some things
5) `yarn performance-test-node`

Output something like this:
```
agorski@zagorski-dev-mee:~/here/harp.gl/mapsdk/coresdk$ yarn performance-test-node
yarn run v1.15.2
$ ts-mocha ./test/performance-tests**/*.ts

...

#performance loading baseline from .profile-helper-baseline.json
#performance OMVDecoderPerformanceTest-theme=berlin tiles=4 region=berlin data=herebase
  min=144.21ms (-7.27% vs 155.51ms) sum=1659.32ms (-5.41% vs 1754.29ms) avg=165.93ms (-5.41% vs 175.43ms) med=156.13ms (-8.99% vs 171.55ms) med95=227.38ms (+8.04% vs 210.46ms) repeats=10 throughput=6.0266/s (+5.72% vs 5.7003/s)**==
#performance OMVDecoderPerformanceTest-theme=berlin tiles=4 region=berlin data=osmbase
  min=148.95ms (-10.22% vs 165.90ms) sum=1645.42ms (-18.23% vs 2012.35ms) avg=164.54ms (-18.23% vs 201.23ms) med=167.10ms (-19.04% vs 206.40ms) med95=182.25ms (-17.31% vs 220.39ms) repeats=10 throughput=6.0775/s (+22.3% vs 4.9693/s)
#performance OMVDecoderPerformanceTest-theme=berlin tiles=10 region=ny data=herebase
  min=590.20ms (-23.5% vs 771.48ms) sum=6643.38ms (-23.44% vs 8677.49ms) avg=664.34ms (-23.44% vs 867.75ms) med=671.41ms (-18.65% vs 825.34ms) med95=714.40ms (-24.75% vs 949.33ms) repeats=10 throughput=1.5053/s (+30.62% vs 1.1524/s)
#performance OMVDecoderPerformanceTest-theme=berlin tiles=10 region=ny data=osmbase
  min=210.74ms (-10.4% vs 235.21ms) sum=2312.38ms (-8.27% vs 2520.75ms) avg=231.24ms (-8.27% vs 252.07ms) med=229.11ms (-7.07% vs 246.54ms) med95=261.45ms (-11.28% vs 294.70ms) repeats=10 throughput=4.3246/s (+9.01% vs 3.9671/s)


```