@nzjony 2019-10-08
The commit message is very detailed, would it make sense to add it to some README.md?