@zbigg 2019-10-09
> > The commit message is very detailed, would it make sense to add it to some README.md?
> 
> Yes please, add it to the README of the package

@nzjony, @musculman Instructions added to main README.