@zbigg 2019-10-10
Merged as part of #853